Rename import back to lib/jsonapi/rails.rb to match gem name
