+++ lib/jsonapi/rails.rb
